utils added to tests
